requirements.txt
anyio==3.6.2
asgiref==3.5.2
click==8.1.3
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
motor==2.5.1
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.3.0
pymongo==3.12.1
python-decouple==3.3
sniffio==1.3.0
starlette==0.16.0
typing_extensions==4.4.0
uvicorn==0.14.0
